# bookstore-back/src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "BookStore APIs exposed from a Jave EE backend",
    "version" : "1.0.0",
    "title" : "BookStore APIs",
    "contact" : {
      "name" : "Lewis Cawthorne",
      "url" : "http://llcawthorne.github.io/",
      "email" : "LLC@acm.org"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/bookstore-back/api",
  "tags" : [ {
    "name" : "Book"
  } ],
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/books" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Returns all the books",
        "description" : "",
        "operationId" : "getBooks",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Books found",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Book resource representation"
              }
            }
          },
          "204" : {
            "description" : "No books found"
          }
        }
      },
      "post" : {
        "tags" : [ "Book" ],
        "summary" : "Creates a book given a JSON Book representation",
        "description" : "",
        "operationId" : "createBook",
        "consumes" : [ "application/json" ],
        "responses" : {
          "201" : {
            "description" : "Book created"
          },
          "415" : {
            "description" : "Format is not JSON"
          }
        }
      }
    },
    "/books/count" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Returns the number of books",
        "description" : "",
        "operationId" : "countBooks",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Number of books found",
            "schema" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "204" : {
            "description" : "No books found"
          }
        }
      }
    },
    "/books/{id}" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Returns a book given an identifier",
        "description" : "",
        "operationId" : "getBook",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "minimum" : 1.0,
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Book found",
            "schema" : {
              "$ref" : "#/definitions/Book resource representation"
            }
          },
          "400" : {
            "description" : "Invalid id"
          },
          "404" : {
            "description" : "Book not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "Book" ],
        "summary" : "Deletes a book given an id",
        "description" : "",
        "operationId" : "deleteBook",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "minimum" : 1.0,
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "Book has been deleted"
          },
          "400" : {
            "description" : "Invalid input.  Id cannot be lower than 1"
          },
          "500" : {
            "description" : "Book not found"
          }
        }
      }
    }
  },
  "definitions" : {
    "Book resource representation" : {
      "type" : "object",
      "required" : [ "isbn", "title", "unitCost" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Identifier"
        },
        "title" : {
          "type" : "string",
          "description" : "Title of the book",
          "minLength" : 1,
          "maxLength" : 200
        },
        "description" : {
          "type" : "string",
          "description" : "Description of the book",
          "minLength" : 1,
          "maxLength" : 10000
        },
        "unitCost" : {
          "type" : "number",
          "format" : "float",
          "description" : "Unit cost of the book",
          "minimum" : 1.0
        },
        "isbn" : {
          "type" : "string",
          "description" : "ISBN of the book",
          "minLength" : 1,
          "maxLength" : 50
        },
        "publicationDate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Publication date of the book"
        },
        "nbOfPages" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Number of pages of the book"
        },
        "imageUrl" : {
          "type" : "string",
          "description" : "URI of image of cover of the book"
        },
        "language" : {
          "type" : "string",
          "description" : "Language book is written in",
          "enum" : [ "ENGLISH", "FRENCH", "SPANISH", "PORTUGUESE", "ITALIAN", "FINISH", "GERMAN", "POTATOHEAD", "RUSSIAN" ]
        }
      }
    }
  }
}